# docs/api/openapi.yaml
openapi: 3.0.0
info:
  contact:
    email: mandarine.app@yandex.ru
    name: Mandarine Support
  description: API for web and mobile application Mandarine
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  title: Mandarine API
  version: 0.0.0
paths:
...
  /v0/account:
    delete:
      description: Request for deleting domain. User must be logged in. User must not
        be deleted.
      operationId: DeleteAccount
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/v0.ErrorOutput"
        "403":
          description: User is blocked or deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/v0.ErrorOutput"
        "404":
          description: Not found user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/v0.ErrorOutput"
        "409":
          description: User is deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/v0.ErrorOutput"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/v0.ErrorOutput"
      security:
        - BearerAuth: []
      summary: Delete service
      tags:
        - Account API
    get:
      description: Request for receiving own domain. User must be logged in. In
        response will be returned own service info.
      operationId: GetAccount
      responses:
        "200":
          description: Account info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/v0.AccountOutput"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/v0.ErrorOutput"
        "403":
          description: User is blocked or deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/v0.ErrorOutput"
        "404":
          description: Not found user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/v0.ErrorOutput"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/v0.ErrorOutput"
      security:
        - BearerAuth: []
      summary: Get service
      tags:
        - Account API
  /v0/account/email:
    patch:
      description: Request for updating email. User must be logged in. In process will
        be sent verification email. In response will be returned updated service
        info.
      operationId: UpdateEmail
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/v0.UpdateEmailInput"
        description: Update email request body
        required: true
      responses:
        "200":
          description: Account info (email is verified)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/v0.AccountOutput"
        "202":
          description: Account info (email is not verified)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/v0.AccountOutput"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/v0.ErrorOutput"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/v0.ErrorOutput"
        "403":
          description: User is blocked or deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/v0.ErrorOutput"
        "404":
          description: Not found user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/v0.ErrorOutput"
        "409":
          description: Duplicate email
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/v0.ErrorOutput"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/v0.ErrorOutput"
      security:
        - BearerAuth: []
      summary: Update email
      tags:
        - Account API
  /v0/account/email/verify:
    post:
      description: Request for verify email. User must be logged in.
      operationId: VerifyEmail
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/v0.VerifyEmailInput"
        description: Verify email request body
        required: true
      responses:
        "200":
          description: OK
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/v0.ErrorOutput"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/v0.ErrorOutput"
        "403":
          description: User is blocked or deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/v0.ErrorOutput"
        "404":
          description: Not found user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/v0.ErrorOutput"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/v0.ErrorOutput"
      security:
        - BearerAuth: []
      summary: Verify email
      tags:
        - Account API
  /v0/account/password:
    patch:
      description: Request for updating password. User must be logged in.
      operationId: UpdatePassword
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/v0.UpdatePasswordInput"
        description: Update password request body
        required: true
      responses:
        "200":
          description: OK
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/v0.ErrorOutput"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/v0.ErrorOutput"
        "403":
          description: User is blocked or deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/v0.ErrorOutput"
        "404":
          description: Not found user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/v0.ErrorOutput"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/v0.ErrorOutput"
      security:
        - BearerAuth: []
      summary: Update password
      tags:
        - Account API
    post:
      description: Request for setting password. User must be logged in.
      operationId: SetPassword
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/v0.SetPasswordInput"
        description: Set password request body
        required: true
      responses:
        "200":
          description: OK
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/v0.ErrorOutput"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/v0.ErrorOutput"
        "403":
          description: User is blocked or deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/v0.ErrorOutput"
        "404":
          description: Not found user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/v0.ErrorOutput"
        "409":
          description: Password is set
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/v0.ErrorOutput"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/v0.ErrorOutput"
      security:
        - BearerAuth: []
      summary: Set password
      tags:
        - Account API
  /v0/account/restore:
    get:
      description: Request for restoring domain. User must be logged in. User must be
        deleted. In response will be returned restored service info.
      operationId: RestoreAccount
      responses:
        "200":
          description: Account info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/v0.AccountOutput"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/v0.ErrorOutput"
        "403":
          description: User is blocked or deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/v0.ErrorOutput"
        "404":
          description: Not found user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/v0.ErrorOutput"
        "409":
          description: User is not deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/v0.ErrorOutput"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/v0.ErrorOutput"
      security:
        - BearerAuth: []
      summary: Restore service
      tags:
        - Account API
  /v0/account/username:
    patch:
      description: Request for updating username. User must be logged in. In response
        will be returned updated service info.
      operationId: UpdateUsername
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/v0.UpdateUsernameInput"
        description: Update username request body
        required: true
      responses:
        "200":
          description: Account info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/v0.AccountOutput"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/v0.ErrorOutput"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/v0.ErrorOutput"
        "403":
          description: User is blocked or deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/v0.ErrorOutput"
        "404":
          description: Not found user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/v0.ErrorOutput"
        "409":
          description: Duplicate username
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/v0.ErrorOutput"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/v0.ErrorOutput"
      security:
        - BearerAuth: []
      summary: Update username
      tags:
        - Account API
...
servers:
  - url: //localhost:8080
components:
  securitySchemes:
    BearerAuth:
      in: header
      name: Authorization
      type: apiKey
  schemas:
...
    v0.AccountOutput:
      properties:
        email:
          format: email
          type: string
        isDeleted:
          type: boolean
        isEmailVerified:
          type: boolean
        isEnabled:
          type: boolean
        isPasswordTemp:
          type: boolean
        username:
          maxLength: 255
          minLength: 1
          type: string
      required:
        - email
        - isDeleted
        - isEmailVerified
        - isEnabled
        - isPasswordTemp
        - username
      type: object
...
    v0.ErrorOutput:
      properties:
        message:
          type: string
        path:
          example: /api/v0/example
          format: url_path
          type: string
        status:
          maximum: 599
          minimum: 400
          type: integer
        timestamp:
          type: string
      required:
        - message
        - path
        - status
        - timestamp
      type: object
...
    v0.SetPasswordInput:
      properties:
        password:
          format: zxcvbn
          type: string
      required:
        - password
      type: object
...
    v0.UpdateEmailInput:
      properties:
        email:
          format: email
          type: string
      required:
        - email
      type: object
...
    v0.UpdatePasswordInput:
      properties:
        newPassword:
          format: zxcvbn
          type: string
        oldPassword:
          type: string
      required:
        - newPassword
        - oldPassword
      type: object
    v0.UpdateUsernameInput:
      properties:
        username:
          format: username
          type: string
      required:
        - username
      type: object
...
    v0.VerifyEmailInput:
      properties:
        email:
          format: email
          type: string
        otp:
          type: string
      required:
        - email
        - otp
      type: object
...